ml_service/requirements.txt
minio==7.2.10
grpcio==1.71.0
grpcio-tools==1.67.0
google-api==0.1.12
googleapis-common-protos==1.65.0
torch==2.5.0
pillow==11.0.0
scikit-image==0.24.0
torchvision==0.20.0
segmentation-models-pytorch==0.3.4
matplotlib==3.9.2
imgaug==0.4.0
numpy<2.0
pydantic-settings==2.6.1
confluent-kafka==2.6.1